tests/kms_big_fb: Fix async
The 'continue' statement causes us to skip test_cleanup()
which means the next iteration of the loop will not re-create
the fb, instead attempting to use the previous fb which has
the wrong format. The result is that things blow up. Skip
the async flip test using igt_require() instead which will let
the cleanup fixture do its job.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>