tests/kms_big_fb: Nuke fliptab[]
Let's just stuff the full transformation into rotations[]
and get rid of fliptab[] and yet another nested loop.

This also extends the hflip tests to be done universally
(as opposed to just in combination with the max-hw-stride
tests).

And as an optimization order the rotations such that the
same fb can be reused between the subtests if possible
(when running multiple subtests in one go).

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>