tests/kms_big_fb: Replace 'bpp' with 'name'
Instead of storing the bpp in the formats[] array let's instead
store a string represtantion. I want to add planar YCbCr tests
to the mix, and so a simple bpp value will no logner cut it.

A bunch of stuff is using the numerical bpp value for some things,
but we can replace those with igt_drm_format_to_bpp().

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>