tests/kms_big_fb: Calculate fb size more uniformly
Unify the max big fb size calculation a bit. For the normal
subtests just stick to the current ~square dimensions, and
for max-hw-stride subtests try to maintain the plane stride
at the maximum value so that we'll actually test the hw limit.

I suppose it's possible the max-hw-stride tests could now
fail if we end up reducing the other dimensions below the
display mode dimensions, but let's cross that bridge if/when
we get to it.

We can also eliminate the special gen3 fence stride codepath
as it's now covered by the max hw stride limites.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>