lib/igt_fb: Make igt_calc_fb_size() somewhat usable
igt_calc_fb_size() is very awkward to use in combinations
with planar/compressed formats. To fix it we either need
to add tons of output parameters (strides/offsets/etc.),
or we just get rid of it and promote calc_fb_size() to
take its place. I chose the latter approach. This does
mean that the callers will need to have a struct igt_fb
around, but I think that's nicer than adding tons of
output parameters.

v2: Fix a misplaced fb->size

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>